Planning the Update Statement
This patch takes the result of the analysis of the UPDATE statement and
generates the necessary query plan. The query plan builds consists of
two nodes: a scan node, that will generate the rows with the values and
key columns that will be updated and a table sink that will process the
incoming row batches and send them to Kudu. Since there are no multi-row
transactions, an error aborting the update process in the middle will
leave some of the rows updated and other in their original state.

Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6998
Tested-by: jenkins

+++ common/thrift/DataSinks.thrift
@@ -56,7 +56,7 @@ struct THdfsTableSink {
 struct TKuduTableSink {
   // The position in this vector is equal to the position in the output
   // expressions of the sink and holds the index into the Kudu schema
-  1: optional list<i64> referenced_columns;
+  1: optional list<i32> referenced_columns;
 }
 
 // Union type of all table sinks.
